ai-chat-branch-be/pyproject.toml
[project]
name = "ai-chat-branch-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.8",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
